refactor: architectury preparation
